requirements.txt
streamlit>=1.32.0
pydantic>=2.7.0
easyocr>=1.7.1
numpy>=1.26.0
Pillow>=10.0.0
python-dotenv>=1.0.0
ollama>=0.1.0

# RAG / Vector Search
langchain-ollama>=0.1.0
langchain-chroma>=0.2.0
chromadb>=0.4.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# User authentication
werkzeug>=3.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0

# Configuration
pydantic-settings>=2.1.0